Website/requirements.txt
flask
firebase-admin
flask-cors
pyjwt
python-dotenv